Refactor SelectClassCommand into directory classcommands
