+++ src/main/java/seedu/address/logic/parser/classcommandparsers/SelectClassCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_CLASS_DISPLAYE
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.SelectClassCommand;
+import seedu.address.logic.commands.classcommands.SelectClassCommand;
 import seedu.address.logic.commands.classcommands.EditClassCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
